directory updates after rebase plus finalize circuit function added
